# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "Documentation for e-commerce APIMy API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get a product",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Product Id",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "number",
                              "example": 2
                            },
                            "name": {
                              "type": "string",
                              "example": "Apple TV 2016"
                            },
                            "imgUrl": {
                              "type": "string",
                              "example": "http://143.42.108.232/products/product-apple-tv.png"
                            },
                            "description": {
                              "type": "string",
                              "example": "The future of television is here."
                            },
                            "price": {
                              "type": "number",
                              "example": 599
                            },
                            "quantity": {
                              "type": "number",
                              "example": 5
                            },
                            "isDeleted": {
                              "type": "boolean",
                              "example": false
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2023-11-28T15:59:06.000Z"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2023-11-30T07:48:26.000Z"
                            },
                            "BrandId": {
                              "type": "number",
                              "example": 1
                            },
                            "CategoryId": {
                              "type": "number",
                              "example": 2
                            },
                            "Category": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "number",
                                  "example": 2
                                },
                                "name": {
                                  "type": "string",
                                  "example": "TVs"
                                },
                                "createdAt": {
                                  "type": "string",
                                  "example": "2023-11-28T15:59:06.000Z"
                                },
                                "updatedAt": {
                                  "type": "string",
                                  "example": "2023-11-28T15:59:06.000Z"
                                }
                              }
                            },
                            "Brand": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "number",
                                  "example": 1
                                },
                                "name": {
                                  "type": "string",
                                  "example": "Apple"
                                },
                                "createdAt": {
                                  "type": "string",
                                  "example": "2023-11-28T15:59:06.000Z"
                                },
                                "updatedAt": {
                                  "type": "string",
                                  "example": "2023-11-28T15:59:06.000Z"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Update a Product",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Product Id",
            "example": 1
          },
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "The Product is being used"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Edit a Product",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "number",
            "description": "Product Id",
            "example": 1
          },
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "iPhone"
                },
                "imgUrl": {
                  "type": "string",
                  "example": "http://143.42.108.232/products/product-iphone.png"
                },
                "description": {
                  "type": "string",
                  "example": "3D Touch. 12MP photos. 4K video."
                },
                "price": {
                  "type": "string",
                  "example": "549"
                },
                "quantity": {
                  "type": "string",
                  "example": "2"
                },
                "isDeleted": {
                  "type": "string",
                  "example": "0"
                },
                "brandId": {
                  "type": "string",
                  "example": "1"
                },
                "categoryId": {
                  "type": "string",
                  "example": "2"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "string",
                          "example": "Product updated"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "All fields are required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/get": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Get all or search products",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "searchOptions": {
                  "type": "object",
                  "properties": {
                    "category": {
                      "type": "string",
                      "example": "1"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 200
                        },
                        "result": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 14
                              },
                              "name": {
                                "type": "string",
                                "example": "Samsung Galaxy Tab A7 Lite (T220) 8.7-inch 16GB Wi-Fi Tablet"
                              },
                              "imgUrl": {
                                "type": "string",
                                "example": "http://143.42.108.232/products/product-samsung-tab-A7.png"
                              },
                              "description": {
                                "type": "string",
                                "example": "Android table - perfect for kids"
                              },
                              "price": {
                                "type": "number",
                                "example": 20
                              },
                              "quantity": {
                                "type": "number",
                                "example": 8
                              },
                              "isDeleted": {
                                "type": "number",
                                "example": 0
                              },
                              "createdAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              },
                              "updatedAt": {
                                "type": "string",
                                "example": "2023-11-28T15:59:06.000Z"
                              },
                              "BrandId": {
                                "type": "number",
                                "example": 2
                              },
                              "brandName": {
                                "type": "string",
                                "example": "Samsung"
                              },
                              "CategoryId": {
                                "type": "number",
                                "example": 1
                              },
                              "categoryName": {
                                "type": "string",
                                "example": "Tablets"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/products/": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Add a Product",
        "produces": [
          "text/json"
        ],
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "iPhone"
                },
                "imgUrl": {
                  "type": "string",
                  "example": "http://143.42.108.232/products/product-iphone.png"
                },
                "description": {
                  "type": "string",
                  "example": "3D Touch. 12MP photos. 4K video."
                },
                "price": {
                  "type": "string",
                  "example": "549"
                },
                "quantity": {
                  "type": "string",
                  "example": "2"
                },
                "isDeleted": {
                  "type": "string",
                  "example": "0"
                },
                "brandId": {
                  "type": "string",
                  "example": "1"
                },
                "categoryId": {
                  "type": "string",
                  "example": "2"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "statusCode": {
                          "type": "number",
                          "example": 201
                        },
                        "result": {
                          "type": "string",
                          "example": "Product added"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "All fields required"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}